requirements.txt
asgiref
black
cfgv
click
colorama
distlib
Django
django-cors-headers
djangorestframework
filelock
geopy
identify
mypy-extensions
nodeenv
numpy
packaging
pandas
pathspec
platformdirs
pre-commit
python-dateutil
python-dotenv
pytz
PyYAML
selenium
six
sqlparse
tomli
typing_extensions
tzdata
virtualenv
pytest
pytest-django
